Fix a bug which caused project and document pages to hang.
The way document file sizes were determined on the project pages were causing PHP to hang. It was performing a HEAD request for the attachment URL to get the Content-Length response. This would usually perform a loopback request back to the web server. However unless the FQDN of the website can be resolved by the docker container, or the URL is rewritten to http://localhost, then the request will hang and block the page from loading.

I've refactored this functionality so that the file size (and now extension) are determined by performing a stat on the local filesystem, rather than a remote HEAD HTTP request. This involved adding a new method get_attachment_id_from_url() to extras.php which is used to get the attachment ID given its URL.